remove the old model_manger and keep refactoring it
