--- src/core/model_manager/CMakeLists.txt
@@ -1,4 +1,0 @@
-set(EXTENSION_SOURCES
-    ${EXTENSION_SOURCES} ${CMAKE_CURRENT_SOURCE_DIR}/model_manager.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/tiktoken.cpp
-    PARENT_SCOPE)
